inmobiliaria_api/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-filter
drf-spectacular
djangorestframework-simplejwt
gunicorn
